requirements.txt
pyelftools==0.28
yara-python==4.2.0
